+++ R/bind_tf_idf2.R
@@ -72,18 +72,18 @@ global_idf3 <- function(sp) {
 
 # entropy
 global_entropy <- function(sp) {
-  p <- sp / Matrix::colSums(sp, na.rm = TRUE)
+  pj <- Matrix::t(sp) / Matrix::colSums(sp, na.rm = TRUE)
   ndocs <- nrow(sp)
-  1 - (-Matrix::colSums((p * log(p)) / log(ndocs), na.rm = TRUE))
+  1 + (Matrix::rowSums((pj * log(pj)), na.rm = TRUE) / log(ndocs))
 }
 
 #' Bind the term frequency and inverse document frequency
 #'
 #' Calculate and bind the term frequency, inverse document frequency,
 #' and tf-idf of the dataset.
-#' This funciton experimentally supports 3 types of term frequencies
+#' This function experimentally supports 3 types of term frequencies
 #' and 4 types of inverse document frequencies,
-#' that are provided in the RMeCab package's functionalities.
+#' which are implemented as RMeCab package features.
 #'
 #' @details
 #' Types of term frequency can be switched with `tf` argument:
@@ -93,10 +93,10 @@ global_entropy <- function(sp) {
 #'
 #' Types of inverse document frequencies can be switched with `idf` argument:
 #' * `idf` is inverse document frequency of which base is 2, with smoothed.
-#' 'smoothed' here means just adding 1 to raw count before logarithmizing.
-#' * `idf2` is global frequency idf.
-#' * `idf3` is probabilistic idf of which base is 2.
-#' * `idf4` is global entropy, not idf in actual.
+#' 'smoothed' here means just adding 1 to raw counts before logarithmizing.
+#' * `idf2` is global frequency IDF.
+#' * `idf3` is probabilistic IDF of which base is 2.
+#' * `idf4` is global entropy, not IDF in actual.
 #'
 #' @param tbl A tidy text dataset with one-row-per-term-per-document.
 #' @param term Column containing terms as string or symbol.
@@ -104,8 +104,13 @@ global_entropy <- function(sp) {
 #' @param n Column containing document-term counts as string or symbol.
 #' @param tf Method for computing term frequency.
 #' @param idf Method for computing inverse document frequency.
-#' @param norm Logical; If true, the raw term frequencies are normalized
-#' being divided with L2 norms before computing values.
+#' @param norm Logical; If true, the raw term counts are normalized
+#' being divided with L2 norms before computing IDF values.
+#' @param rmecab_compat Logical; If true, computes values while
+#' taking care of compatibility with RMeCab.
+#' Note that RMeCab always computes IDF values using term frequency
+#' rather than raw term counts, and thus TF-IDF values may be
+#' doubly affected by term frequency.
 #' @return data.frame.
 #' @export
 #' @examples
@@ -127,18 +132,14 @@ bind_tf_idf2 <- function(tbl,
                          n = "n",
                          tf = c("tf", "tf2", "tf3"),
                          idf = c("idf", "idf2", "idf3", "idf4"),
-                         norm = FALSE) {
+                         norm = FALSE,
+                         rmecab_compat = TRUE) {
   tf <- rlang::arg_match(tf)
   idf <- rlang::arg_match(idf)
 
   term <- as_name(ensym(term))
   document <- as_name(ensym(document))
   n_col <- as_name(ensym(n))
-
-  sp <- cast_sparse(tbl, !!document, !!term, !!n_col)
-  if (isTRUE(norm)) {
-    sp <- Matrix::t(Matrix::t(sp) * (1 / sqrt(Matrix::rowSums((sp * sp)))))
-  }
 
   terms <- as.character(dplyr::pull(tbl, term))
   documents <- as.character(dplyr::pull(tbl, document))
@@ -154,19 +155,27 @@ bind_tf_idf2 <- function(tbl,
       )
     }
   )
+  if (identical(tf, "tf")) {
+    tbl <- dplyr::mutate(tbl, tf = .data$n / as.numeric(doc_totals[documents]))
+  } else {
+    tbl <- dplyr::mutate(tbl, tf = purrr::flatten_dbl(doc_totals))
+  }
 
+  if (isTRUE(rmecab_compat)) {
+    sp <- cast_sparse(tbl, !!document, !!term, "tf")
+  } else {
+    sp <- cast_sparse(tbl, !!document, !!term, !!n_col)
+  }
+
+  if (isTRUE(norm)) {
+    sp <- Matrix::t(Matrix::t(sp) * (1 / sqrt(Matrix::rowSums((sp * sp)))))
+  }
   idf <- switch(idf,
     idf = global_idf(sp),
     idf2 = global_idf2(sp),
     idf3 = global_idf3(sp),
     idf4 = global_entropy(sp)
   )
-
-  if (identical(tf, "tf")) {
-    tbl <- dplyr::mutate(tbl, tf = .data$n / as.numeric(doc_totals[documents]))
-  } else {
-    tbl <- dplyr::mutate(tbl, tf = purrr::flatten_dbl(doc_totals))
-  }
   tbl <- dplyr::mutate(tbl,
     idf = as.numeric(idf[terms]),
     tf_idf = .data$tf * .data$idf